InstantScores refactor + Int to Long changes (#1337)
* InstantScores refactor + Int to Long changes

* Fix lint

+++ app/src/main/java/com/jerboa/model/SiteViewModel.kt
@@ -106,7 +106,7 @@ class SiteViewModel(private val accountRepository: AccountRepository) : ViewMode
         }
     }
 
-    private fun getUnreadCountTotal(unreadCountRes: ApiState<GetUnreadCountResponse>): Int {
+    private fun getUnreadCountTotal(unreadCountRes: ApiState<GetUnreadCountResponse>): Long {
         return when (val res = unreadCountRes) {
             is ApiState.Success -> {
                 val unreads = res.data
